added changes from godtools-ios. they were accidentally added there and the history behind this commit can be found prior to this tag GTViewController-pod-history which can be viewed at https://github.com/CruGlobal/godtools-ios/tree/GTViewController-pod-history

+++ GTViewController/Classes/Custom Classes/UISnuffleButton.m
@@ -99,7 +99,7 @@ extern NSString * const kButtonMode_allurl;
         NSTimeInterval currentTimeStamp = [touch timestamp];
 
 		//calculate changes in direction and time since last call to touchesMoved
-        double deltaX                   = fabs(currentTouchPosition.x - self.lastTouchPosition.x);
+        double deltaX                   = fabsf(currentTouchPosition.x - self.lastTouchPosition.x);
         double deltaT                   = currentTimeStamp - self.lastTimeStamp;
 
 		//add the current change in distance to the total distance
